refactor: separated and moved roomDB to data.api package
